requirements.txt
asgiref==3.4.1
dj-config-url==0.1.1
dj-database-url==0.1.2
Django==4.0.1
gunicorn==20.1.0
sqlparse==0.4.2
tzdata==2021.5
whitenoise==5.3.0
psycopg2==2.8.6
djangorestframework
djoser
python-decouple==3.1
django-cors-headers==3.11.0
django-filter==21.1
drf-extra-fields==3.2.1